BAU: Allow tests to be run in a merge queue
We're going to try out merge queues[^1] as a solution for constantly needing
to rebase our PRs when another PR is merged.

Some of the steps in this workflow are required before a PR is merged.
The same restriction applies to Github merging a PR from a merge queue,
so before we can set up the queue we need to allow the workflow to be
triggered by the queue.

[^1]: https://docs.github.com/en/repositories/configuring-branches-and-merges-in-your-repository/configuring-pull-request-merges/managing-a-merge-queue#about-merge-queues