OLH-1766: Enable the Cloudfront header in all environments
We've confirmed with TxMA that they're ready to receive events with this
new field in production, and with our service transition lead that this
is a minor change for us since we're not doing any of the DNS changes.

That means we're ready to call the function in production and start
setting the new header for requests to our frontend.

Remove the condition since we want this on in every environment.