OLH-1804: Update contact hours
Remove the option to contact GOVUK OL on the weekend as this will no
longer be available.
For the time being we are simply removing the weekend bullet point,
until we get back Welsh translations for the paragraph version (i.e.
"You can call us Monday to Friday, 8am to 8pm" instead of "You can call
us on: • Monday to Friday, 8am to 8pm". This is because it's fairly
urgent that the content is updated asap.