OLH-1835: update phone/webchat opening hours
Change opening hours for phone and webchat to a paragraph format as
opposed to a bullet list format.
The option to contact the contact centre on the weekend has been
removed. We have now received the Welsh translations so the content can
be re-formatted as a sentence.

Also, remove random empty <p> tag.