BAU: Add file extensions to all imports
These are now required in an ESM module. Confusingly, they have to be
`.js` instead of the `.ts` we might expect. That's because Typescript
doesn't rewrite import paths when it complies to JS.

You can read some of the many issues discussing this if you want to know
why (eg. https://github.com/microsoft/TypeScript/issues/42151)