BAU: a11y statement inset text tweak
Tweak the way the webchat accessibility statement paragraph is displayed
on the contact triage page.
The inset text used to have several paragraphs which were stored as an
array in the translation files. This has since been updated so there is
one paragraph only, which means the current solution is a bit
overengineered.
Some small style updates: use em dash instead of a regular hyphen and
move the full stop outside of the anchor tag.