OLH-1868: Update wording around "security codes"
This update is to reflect Auth's content changes, removing the phrase
"security code" and replacing it with simply "code".

The reason behind this update is that some mobile browsers interpret
input fields labelled "security code" as being credit card related, and
therefore prompt users to enter their card security code which is
obviously not correct.