INCIDEN-834: Include new phone number in payload
It looks like when we refactored this controller to switch between the
new MFA method API or the old API based on the feature flag we set up a
common type definition for both. In #1401 we removed the new phone
number from the payload when we added new logic for the endpoint field,
but this inadvertently also removed it for the old method which is still
in use at the moment.

This means users cannot currently change their phone number. Add the new
value back into the payload for all requests.

After this PR we'll need to remove this new field from the payload in
the new-style API calls, but that can be done in a followup PR since
it's not in use yet.