BAU: Refactor the container configuration to a mapping
Refactor the container CPU, memory, min containers and max containers
out into a mapping per-environment.

We're going to want to tweak these values in response to performance
testing, and we'll probably end up on different values in different
environments to save money in eg. staging where there's very little
traffic.

There's no changes to these values in any environment at the moment, but
this will make it easier to do so when we want to.