OLH-1991: Add an initial post-deploy test
This uses Jest, jest-cucumber and Puppeteer to run a test which loads
the contact page and checks the status code is 200.

Having this as a post-deploy test allows us to at least confirm the app
has started and is able to serve requests, and jest-cucumber lets us
structure our tests in a sensible way to build on in the future.