OLH-1991: Run post-deploy tests in a Docker container
Set up a Docker container to run the post-deploy tests. This uses the
Debian based Node image because alpine has the musl libc compiler which
can't compile the AWS CLI.

We also need to install Chrome separately because Puppeteer's default
install of Chromium doesn't work since the base image doesn't include
all the necessary shared libraries.

This does mean we can *only* run these tests inside the Docker container
now. It should be fairly simple to allow us to switch the config if
developers want to run them on their machines without Docker, but that's
out of scope for this initial setup.