OLH-1991: Add a Github action to build the test image
Add a new Github action that builds and pushes the post-deploy tests
image to ECR in our build environment. I've already set up the ECR
repository as described in the documentation[^1].

This action is triggered when a PR is merged to main that includes
changes to the post-deploy tests folder or the action itself. I've set
this up as a separate action rather than a new step in the existing
post-merge actions so it doesn't add to the time between merging code
and it getting deployed to production.

The image needs to be pushed to ECR before the pipeline starts running
the test step, but I think we should be safe to do this as a parallel
step. The only slow part in this action is building the image, most of
which is network time downloading Chrome. It takes well under 5 minutes
on my laptop and the Github action runners have faster internet
connections than me.

This is comfortably under the ~7 minutes it takes for Github actions to
test, build and push the main frontend image plus ~7 minutes for the
stack to update in Codebuild.

[^1]: https://govukverify.atlassian.net/wiki/spaces/PLAT/pages/3107258369/How+to+deploy+a+container+to+Fargate+with+secure+pipelines#Step-9%3A-Configure-pipelines-for-running-tests-against-infrastructure-(Optional)