OLH-1991: Downgrade image upload action versions
This action is currently failing with a permissions error:

```
denied: User: arn:aws:sts::301577035144:assumed-role/account-mgmt-frontend-pipeline-GitHubActionsRole-AHD0ZJ9UGXU4/GitHubActions is not authorized to perform: ecr:InitiateLayerUpload on resource: arn:aws:ecr:eu-west-2:301577035144:repository/*** because no identity-based policy allows the ecr:InitiateLayerUpload action

```

I think this is because the newer version of one of these actions is a
breaking change and slightly changes the role ARN that we end up using
(`assumed-role` in the ARN instead of just `role`).

The action to upload the image to our dev environment does almost
exactly the same steps and that works, so downgrade the versions in this
action to match it.