OLH-1991: Disable interactive pager for test image
By default the AWS CLI uses `less` to display an interactive pager on
commands that have a long expected runtime (eg. uploading a file to S3).

This Docker image doesn't have `less` installed, so the tests were
failing when the pipeline ran the AWS command to get the frontend stack
outputs.

Disable the interactive pager so the command can run in the headless
container