OLH-1263: Overwrite govuk tag with old styling
There have been concerns that different frontend applications within OL
upgrading to govuk-frontend v5+ on different timelines would result in a
degraded UI experience due to the visual difference in the tag component
(and consequently, in the phase banner) which is implemented in v5.0.0.

This commit aims to work around this by overruling the new CSS for
govuk-tag with legacy CSS that precedes govuk-frontend v5.0.0.
This change is intended to be temporary. We should migrate to the new
phase banner style once all the other frontends in OL have upgraded to
the newer versions of govuk-frontend.