OLH-2207: Fix translation errors
Logs in cloudwatch were showing multiple errors flagging that key
'/track-and-redirect' was missing for the en and cy languages.
There was no such key in our translation files. I traced it back to the
contact triage page where the nunjucks translate filter was accidentally
used on a local variable. This didn't have any user-facing consequence
so it went unnoticed for a while.

I addressed a couple other errors at the same time:
- on the your services page, it seems like the details element was
  removed but the njk html block that used to be its content was not.
The contents of that nunjucks block were leading to a translation error
as the relevant strings had been removed from the translation files
- on the security page, the translation being passed as visually hidden
  text for the phone number "change" link was incorrect