BAU: remove redundant call to showPassword script
The showPassword script has been removed as part of the work to replace
the OLH local show password component with the DS equivalent.
I missed this <script> tag in the process of removing the old component.
All assets related to this component are now automatically imported from
the DS package (govuk-frontend)