OLH-2432: content tweaks on activity log
Make some content (mostly) related tweaks on the activity log page, for
the "RSA off" scenario.
- remove paragraph before the dropdown as it is being duplicated in the
  dropdown again
- "change your password" link should not open in a new tab
- GOV.UK One Login was misspelled

One small change that is semi-related is I moved the block of content
for the security notice dropdown into the njk template as I couldn't see
any particular reason why it should be in the controller.

Additionally, added Welsh translations for the new content.