OLH-2526: Reword RSA allow list to Activity Log
The original plan was that activity log and the (OLH owned) RSA journey
would go hand in hand – there would be an allow list of services whose
users would all have the ability to access the RSA journey.

In the meanwhile the requirement has shifted so there will be a rollout
of the activity log with a "new" RSA capability (in the shape of a form
developed by the auth team). The "old" RSA journeys (developed by OLH)
are to remain switched off for now.

There is now a requirement to be able to toggle these three capabilities
on/off independently (the activity history view, the "old" OLH RSA
journeys and the "new" RSA form).

In light of this, the terminology around "RSA allow list" has been
updated in our code to reflect the updated requirement – the activity
log is essentially being rolled out with a RSA-lite where instead of
each item in the activity history having its own Report link, there is a
catch-all contact link at the top of the page.
The allowlist-ing is of services whose users get to access the activity
history, but these users will not be able to RSA (not via the journeys
previously developed by the OLH team).