OLH-2526: Update reporting info paragraph
Update content on the reporting info paragraph at the top of the
activity history page.
There has been a decision to introduce a separate flag to allow us to
switch the ability to navigate to the auth RSA form on/off where needed
(i.e. when support requests exceed current capacity)

When the flag is on, the content will appear as a dropdown with a link
to the form, and a paragraph telling users to change their password.
When the flag is off, the content will appear as one paragraph advising
users to change their password.