OLH-2140: Use AuthApp instead of AuthenticatorApp in URL
This is consistent with the rest of the content.

I could have just changed the key since that's the only part which
shows up in the URL, but all the other states have have the state
machine variant, key etc. all called the same thing. I think we'd be
making our future lives harder if we made this one different.