OLH-2140: Display a backlink on the change auth app page
Display a backlink on the change authenticator app page which takes you
back to enter password with the correct `type` for this journey. This
works for both the GET and POST responses when the template is
re-rendered with errors.

I'm 90% sure this works fine with our state machine. I've clicked back
and forth a few times and I'm able to complete the journey.