OLH-2534: Set up the base for a new MFA API client
I find the existing logic / functions we have for this API really
complex and hard to understand. I think that's because they've ended up
with many abstraction layers and passing large config objects down
through those layers.

I think we ended up *there* by working on this feature top down from the
designs, and doing each journey one by one. It means the logic is split
across many components' services and we don't have a single interface to
the API that exposes simple types etc. Our tests are similarly split up
which makes it difficult to reference the expected request / response
objects to the API spec.

I believe we can remove a lot of the complexity by building our API
client 'bottom up' based on the spec.

The information we're passing in the config objects is all request
scoped and known in the controller, so we can initialise an API client
class at that point and then only pass the relevant information when we
use the client to make API calls.