OLH-2534: Handle error responses in the MFA API
Wrap the API response in an object that includes the response data,
status code, if the request was a success and any problems.

This structure should make it easier for calling code to understand if
an operation completed successfully without needing to know the details
of what makes a successful / unsuccessful response.

Notably the API will respond with either a 200 or 204 status for a
successful request depending on the method we're calling (204 for
DELETEs).