OLH-2543: Define static routes first
The order of routes and middleware in `app.js` matters. Incoming
requests are matched and middleware is run in the order in which it's
initialised [^1]. We need the middleware, i18n, local variables etc. to be
available for requests to our application pages but we don't need any of
that for the healthcheck or our static CSS & JS.

Notably this moves the static route definitions before we initialise our
session store client. I'm not sure if creating the client involves a
call to DynamoDB but it's still a lot of setup we don't need for
static asset requests.

Hopefully we'll be able to see a difference when we next run the
performance tests.

[^1]: https://team-manual.account.gov.uk/development-standards-processes/coding-practices-and-processes/express/#app-js-ordering