OLH-2534: Use new API client in delete MFA journey
Switch the delete MFA method journey to use the new API client added
in #2068. This doesn't bring us a lot of complexity savings just yet but
I picked this journey to start with as it's the simplest.

The repeated `as unknown as MfaMethod` type casts and `NewMfaMethod` in
the session type definition aren't ideal but they're needed to keep the
existing code working on the other journeys. I will clean them up
afterwards once I get to the end.