OLH-2534: Throw an error if deleting a user's MFA fails
This error will be picked up by our middleware and the user will see the
'Sorry, something went wrong' page.

I think it's clearer to explicitly perform this check at the point where
the response comes back in the controller and then throw a new error,
rather than rely on the API client or axios to throw one.

This way means we can add a unit test at the controller too.