OLH-2534: Add a function to create an MfaClient
We'll need to instantiate a client in each controller that uses this
API. The constructor is a little chunky due to all the IDs we need to
pass on every request to the API.

All of those IDs are present in various places on the `req` and `res`
objects, so this refactor adds a new function that takes those objects
and returns an instance of an `MfaClient`.

This will reduce the lines of code and repetition in each controller. It
also has the benefit of reducing the mock complexity in controller tests
- we can now stub this function to return a `sinon.SinonStubbedInstance<MfaClient>`
and we'll get an object that has all the type definitions from the base
class but that allows us to set our own return values for all the
methods.