OLH-2534: Increase test coverage for adding an app MFA
Add some simple tests to make sure the controller is re-rendering the
'Set up an authenticator app' page if one of the validation conditions
fails. These don't check for the specific error message on each variant
but they should give us more confidence that the app is generally doing
what we think it is.