OLH-2542: Update to the latest client registry
This marks all the English only RPs as such. See https://github.com/govuk-one-login/di-account-management-rp-registry/pull/44

This has a side effect of changing the example RPs we use for our
services test. If these change again we should stub the RP registry in
this test file but I don't see that happening in the near future.