OLH-2534: Add a makeDefault method to the MFA API client
It's possible that we'll need to make an MFA method default from more
than one controller, so it'll save us some repetition if we put the
logic in the API client.

It also separates the concerns of the controller (make the method
default) and the API (what request format do I need to make this call).