OLH-2534: Refactor the MFA method types
When I initially set this up I thought that the Typescript compiler
would allow us to pass any interface that extended `Method` to a
function that required a `Method` and then would infer which interface
it was based on the properties.

Unfortunately that's not quite how it works. The more Typescript-native
approach to allowing multiple options for the shape of an object is a
union type. This achieves the original functionality I wanted and also
gives some extra protections as when we use the union type the compiler
forces us to write code that can handle each variant.