OLH-2534: Only use the API client's MFAMethod type
This replaces all use of the previous `MfaMethod` type with the new one
exported by the API client everywhere except the old library. Crucially
it updates the session definition so I can remove all the ugly `as
unknown as MfaMethod` type casts I needed before.

This is safe to do so because there's sufficient overlap between the two
types since we updated the old type to use a string as the identifier[^1].

[^1]: https://github.com/govuk-one-login/di-account-management-frontend/pull/2052