OLH-2534: Pass the phone OTP when creating an SMS method
When a user sets up a new auth app method we do all the OTP validation
on our side and simply pass the API the secret credential.

However when a user sets up an SMS method we send them an OTP via text.
Our frontend can't validate that OTP as it's managed by the separate
account management API. Instead we need to pass it in our API call when
we try to create the new method.

This commits adds an optional `otp` parameter to the `MfaClient.create`
method. That meant we needed some validation about the combination of
OTP / method types and the API client method was getting quite clunky so
I moved the validation out to a separate file.