OLH-2534: Update the error type returned by the MFA API
This now matches what's returned by the API.

Originally the API design was going to use RFC 9457[^1] as a standard
format for error messages returned by the API. We'd set up the parsing,
interfaces etc. to handle errors in this format.

We've now changed our mind about using the format described in that RFC
and gone with a simpler error structure.

I've also added a helper function that takes a response and formats the
error data into a nice string. I considered making a new class that
implements the `ApiResponse` interface so that function could be a class
method but it felt like a lot of boilerplate for what's essentially the
same outcome. If we have one more function that needs to operate on a
response object we should reconsider that decision.

[^1]: https://www.rfc-editor.org/rfc/rfc9457.html