OLH-2534: Use MFA API client when adding a new SMS method
This is the first change that touches the SMS OTP journey. There's a lot
going on in that controller and this commit only adds to the complexity
with some more nested if / else blocks.

I'll be able to remove most of these once I come back here to make the
same changes on the 'update phone number' journey.