OLH-2534: Create new MFA methods as backup
Users can't create new MFA methods as default - they'll always need to
create it as a backup and then switch their backup and default.

This fixes a bug we saw in staging where I was unable to create a new
backup auth app method. Even though I was going through the journey to
create a backup method the API client was always trying to create it as
default, which was correctly rejected by the API.