OLH-2534: Log the API error message when adding a method
This will make it easier to identify the source of issues in the future.

The previous approach threw an error with a useful message, but our top
level error handler catches these and only logs the failed status code.
That's because the stack trace for an unhandled error might contain
personal data which we can't store in our logs.

The API error codes and messages won't contain any personal data, so
we're safe to log a more useful error message before throwing the error.