OLH-2560: Use Firewall Manager in build and staging
We've successfully tested in dev so we can simplify the template and
make the tagging changes at the same time as we disassociate our custom
ACL.

There is a risk of a few seconds downtime between us removing our ACL,
the tags being applied and Firewall Manager taking over. This has been
investigate and accepted by security.