OLH-2534: Allow passing an OTP to the update API method
The rules for this are a little complex. If we are changing a phone
number on a default method or switching an existing method from app to
SMS then we need to also provide the OTP that's sent to the user's new
phone number.

However if we are switching a backup to default then this also uses the
update API endpoint, but doesn't need the code as both phone numbers
have already been verified.

I've tried to capture this in the validation rule but it's not able to
catch evety case as the API client doesn't know the context in which
it's making the update. I thought this was enough to prevent most
mistakes while maintaining separation of concerns.