OLH-2534: Use the MFA API client when a user changes their default method
This journey won't work in staging yet as the API doesn't currently
permit you to change the type of an existing MFA method (eg. SMS -> auth
app). I've been speaking to the Auth team about this and they agree it's
a sensible change to make.

There is a possible workaround where instead we make multiple API calls
to:

1. switch the default and backup methods
2. delete the new backup
3. create a new method with the details the user's entered
4. switch the backup and default again to make the new one the default

This is risky however as the group of API calls isn't transactional, so
an error halfway through could put the user in an unknown, or worse,
unrecoverable state. This feature isn't live yet so it's not worth the
effort to implement it with the multiple API calls if we know we're
going to change the API to support the simpler way.