OLH-2534: Refactor check your phone controller
Moves the MFA handling logic back out into a separate function to reduce
complexity in the main controller. I've also gone with a switch / case
statement as we're operating over the different intents - I thought this
was easier to follow than several if / else if / else blocks.

Having all the MFA logic in one place also means we can define the
reponse at the start and de-duplicate all the error handling etc.