OLH-2557: Use the new headers function up to integration
There was an issue with some AWS accounts not being part of the correct
Oragnisational Unit in Control Tower which meant the dev platform team
had to deploy a new version of the headers function to update it with
the new device intelligence requirements.

Update our distribution to point to this new function in all
environments except production. I'll then let the dev platform team know
and they can verify the function is being called.

Once we get the ok there, I'll release to prod.