BAU: Do linting validation on the Cloudformation template
There's a new(ish) option to `sam validate` which calls out to
`cfn-lint` to actually check the structure of the template and all the
resources, rather than the basic checks you get without the option.

We want to catch any errors early, before we deploy, and this is another
tool to help us do that.