OLH-2584: Don't show offboarded services as cards
or on the 'delete your account page'.

Both these pages filter the list of services a user's accessed using the
list of account and service RPs from the registry.

The activity log doesn't do this filter - it just gets all the entries
for that user from the table.

This means filtering these two lists will stop offboarded services from
appearing in service cards or on the 'delete your account' page, but
they'll still show and have the correct content in users' activity
history.