OLH-2592: Change the payload when switching an MFA method to default
The Auth team have decided to update the required payload when switching
an MFA mthod from backup to default. This reduces the number of fields
and will make it easier for them to validate the request.

The new format is
```
PUT /{publicSubjectId}/{mfaIdentifier}
{
  "mfaMethod": {
    "priorityIdentifier": "DEFAULT"
  }
}
```

Update the API client and the controller where we call this method.