OLH-2603: Nunjucks global flag and CSS tweak
Implement global flag using Nunjucks as per the guidance in the release
notes.

Tweak CSS for contact page:
The alignment of the GOVUK OL link and Sign out link on the contact page
is off in the context of the new styling for the refreshed GOVUK brand.
This helps the product name and sign out link remain in line with each
other on smaller viewports.