OLH-2546: Update Webchat URLs for dev & UAT
The supplier for the webchat on the contact page has created a
designated infrastructure to decouple us from other customers of the
webchat.
The supplier have requested confirmation that the UAT version of the
webchat works before they allow testing of prod webchat in our
integration environment.

This updates the source urls for the dev and uat versions of the webchat
which will enable us to test the new src  in these environments.