OLH-2642: Replace custom nav bar with DS component
The secondary navigation bar on GOVUK One Login home precedes the Design
System Secondary Navigation component which was more or less created in
its image.

This replaces the custom implementation of the secondary nav on OLH with
the DS component, noting some differences in behaviour, namely in the
fact that the DS component collapses into a dropdown menu on mobile.