OLH-1704: Replace `csurf` with `csrf-sync`
This replaces our CSRF protection library as `csurf` has been
deprecated.

`csrf-sync` uses the [Synchroniser Token Pattern](https://cheatsheetseries.owasp.org/cheatsheets/Cross-Site_Request_Forgery_Prevention_Cheat_Sheet.html#synchronizer-token-pattern)
to do protection. It's mostly a drop in replacement for us.

It works by storing a token in the user's session. That's made available
to the templates through `res.locals`. Any template that has a form has
a hidden `_csrf` field which gets populated with that token. When the
form is submitted the middleware checks for the token and validates that
it matches the one in the session.

All that is exactly the same behaviour as before. The only difference is
where the token is stored - in the session not in a cookie.

This is using one token per session, not one per request as can be
configured. I tested using one per request but I was able to pretty easily get it to
fail with aborting requests and then refreshing, It seems that I'd got
through enough processing to generate a new token, but not enough for it
to get saved into the DynamoDB session store by the time I ended the
request. This will probably also be an issue with double submits or very
slow page loads.

We had one token per session in the old implementation, so I don't see
any issues with the same config here.