BAU: Replace deprecated `getNextSnapshot` with `transition`
`getNextSnapshot` returns a `SnapshotFrom<T>` and `transition` returns a
tuple where the first value is the same `SnapshotFrom<T>`. We don't need
the second item in the tuple (an array of actions) so we can replace the
deprecated function and destructure the array to get the value we need.

See https://www.jsdocs.io/package/xstate#getNextSnapshot