OLH-2680: Make it clear array is sorted in place
`Array.sort()` sorts the original array in place and also returns a
reference to the same array. Assigning this reference to a new variable
can be confusing if we ever needed to do anything else with the original
variable, as it implies that the original hasn't been sorted.

There's no issues with that in the case, but we can make our intentions
clear by not assigning the result of `.sort()` and returning the
original variable.