OLH-2642: Navigation tweaks
Make internal OL Home navigation match the DS secondary navigation
better from a visual perspective, while maintaining its behaviour on
mobile.
After discussions with the UCD team, it was determined that the custom
secondary nav in OLH shouldn't be swapped out for the DS service
navigation component as the styling we currently have suits our needs
better (specifically, the custom nav on OLH should not turn into a
dropdown menu on mobile, instead the links should remain inline;
additionally, we've determined the larger font size we are currently
using is more user friendly).

We decided to make some tweaks to current styling to make it look more
similar to the DS component, namely:
- tweaking font weight
- ensuring background colour changes with the rebrand.