OLH-2691: Use an enum for audit event names
We only want to allow the app to send audit events with names on a
pre-defined list. Previously we had an object that mapped an allowed key
to a string, but the functions that build and send audit events accepted
any string as the event name.

Swtiching this over to an enum means the type system will force us to
only provide event names from the enum values and (I think) makes the
intent clearer.