OLH-2691: Use `const enum`
Enums are a controversial feature in Typescript. There seem to be two
main reasons people don't like enums:

1. The map `as const` is probably marginally faster at run time than an
enum as it can be transpiled into literals instead of an object. I can't
see anywhere with a concrete benchmark on that and I don't think it's
very important in this case.

2. If you allow Typescript to generate values for your enum and then
change the order of the variants the values for each will change. Again
that's not an issue here as we'll always provide a string value for each
variant.

3. `const`s are readonly so you can't add a value at runtime. We don't do
anything like that so this would be nice to have.

The good news is that because we're also not using any of the enum
features like iterating over the variants we can have the best of both
worlds with a `const enum` 🎉 . Those get the same syntax and
typechecking behaviour in your editor but are transpiled to literals at
build time so are also as fast as an `as const` object and prevent
runtime updates.