OLH-2691: Refactor event service to allow different data
Currently the event service builds all audit events with the same fields
and data. That made sense when we only had events related to the contact
page, but we're about to add several new ones for the MFA work and
they'll need different fields and logic.

This structure allows for that while keeping the same simple interface
for the service's API. If the logic in each branch of the statement
starts to get very complex we can move it all out into a separate
builder class, but we're not at that point yet.