OLH-2783: Use a session scoped session cookie
Remove the max age definition from our session cookie config. This means
express-session won't set an `Expires` field on the cookie and therefore
browsers will treat our session cookie as scoped to that browser
session.

In practice, this means that we'll log users out when they close the last
tab / window with Home open in it.