OLH-2506: Remove q param from English page link
On the searchable list page, when there are no results after searching
for a service in Welsh, users are directed to search the English list
instead.

It has been advised that in this scenario, we should link to the reset
version of the English language page, as users in this scenario would
likely be looking to re-try the search in English anyway.