OLH-2506: change input type from search to text
Changing the type of input means that a "X" no longer appears inside the
text field. This is on purpose as having the inline X button (which
clears the field but doesn't re-submit the search) in addition to the
"Clear search" link could prove confusing to users.

I've investigated the potential accessibility implications but the only
difference seems to be in the way the field is announced to screen
readers ("text" vs "search"). The form still appears in the landmarks
list as a `search` landmark and functions the same way.