OLH-2802: Rename global logout to backchannel logout
This is for 2 reasons:

1. In the OIDC spec and our documentation[^1] this feature is called
   backchannel logout. Giving our component the same name as everywhere
   else will make it easier to understand what's going on.

2. We're about to add a new feature called 'global logout' which will
   allow users to force a logout from all sessions including the GOV.UK
   app. Having two components with the same name, or a feature with a
   different name to the component will be a recipe for disaster.

This renames the component and adds a new route for it. I've also kept
the old route for now. I'll need to ask the Orchestration team to update
their registry with the new URL before I can get rid of the old one.
Once that's done I'll remove the old route and entry in the constants
file in a separate PR.

Having both route available at the same time means we can safely merge
this PR without needing to coordinate an update with the Orchestration
team.

[^1]: https://docs.sign-in.service.gov.uk/integrate-with-integration-environment/managing-your-users-sessions/#request-logout-notifications-from-gov-uk-one-login