BAU: Update ML websocket addr for integration env
A while ago the webchat source URL for integration was updated to the
new production URL. This was done in order to test the new
infrastructure in integration before proceeding to prod.

It was missed that the websocket URL also needed updating.

This updates the relevant env variable in order to enable connection to
the prod websocket on our integration env. This (along with the Webchat
src URL change) should be reverted as soon as testing is complete.