BAU: Only use `APP_ENV` to determine environment
Setting `NODE_ENV` to anything other than `production` is considered an
anti-pattern and is discouraged[^1].

This is to prevent issues where we write code that behaves differently
in our tests vs. when deployed.

That means we should also not use this variable to determine what
environment we're currently in. This replaces any use of `getNodeEnv`
with the more appropriate `getAppEnv` and removes `getNodeEnv`.

I've also updated the `ENVIRONMENT_NAME` constant to:
  1. a const enum
  2. actually reflect the environment names, not the NODE_ENV options

This will make it easier for us to understand the intent of switches for
different behaviour in environments.

[^1]: https://nodejs.org/en/learn/getting-started/nodejs-the-difference-between-development-and-production