OLH-2946: Remove global try catch
Remove the `globalTryCatch` and `globalTryCatchAsync` functions. These
wrap our controller functions and catch unexpected exceptions, increment
a metric and then pass on to the next middleware.

We also have a log error middleware which does the same thing, so we've
got duplicate logic. The middleware is a better approach as Express will
run it on all requests / controllers and means we can't forget to add
the wrapping function.

We think these functions were added back when we were on an old version
of Express which didn't handle errors in async controllers. Now we're on
Express 5, `next()` is automatically called with a rejection if the
controller returns a Promise.