BAU: add skipTarget-local for webchat test
The webchat test is failing for some of us locally although it is
passing in CI. This is probably to do with the websocket connection
errors we have been experiencing for a while.

This disables webchat test locally. Tests will still be running in theGH
action. Also added @postDeploy and removed the cookie step as per
Michael's suggestion.

The cookie step is unnecessary as the contact page functions the same
whether cookies have been accepted or not, and there is a separate
cookie consent banner test.